Implemented offline-first functionality.
Bug: back gesture inside expanded search banner doesn't work after getting back from settings screen.

+++ app/src/main/java/com/enjot/materialweather/domain/model/HourlyWeather.kt
@@ -6,7 +6,7 @@ data class HourlyWeather(
     val localFormattedTime: String,
     val temp: Int,
     val humidity: Int,
-    val windSpeed: Double,
+    val windSpeed: Int,
     val windDeg: Int,
     val pop: Double,
     val rainPrecipitation: Double?,
